refactor: main.basic.js 관심사 분리
main.basic.js 파일의 가독성과 유지보수성을 향상시키기 위해 관심사 분리 원칙에 따라 리팩토링을 진행했습니다.

주요 변경 사항:

   * 기존의 거대한 createShoppingCart 함수를 다음과 같은 모듈로 분리했습니다.
       * constants.js: 상수 모음
       * state.js: 상태 객체
       * calculation.js: 순수 계산 함수
       * dom.js: DOM 생성 및 렌더링
       * handlers.js: 이벤트 핸들러
       * utils.js: 공통 유틸리티 함수 ($)
   * main.basic.js는 각 모듈을 가져와 초기화하는 역할만 담당하도록 수정했습니다.
