+++ src/basic/utils.js
@@ -1,0 +1,1 @@
+export const $ = sel => document.querySelector(sel);
